Implement image texture

+++ RayTrace/Perlin.h
@@ -51,9 +51,9 @@ public:
 		float v = p.y - floor(p.y);
 		float w = p.z - floor(p.z);
 
-		int i = floor(p.x);
-		int j = floor(p.y);
-		int k = floor(p.z);
+		int i = int(floor(p.x));
+		int j = int(floor(p.y));
+		int k = int(floor(p.z));
 
 		glm::vec3 c[8];
